spdm-lite: add buffered large response chunking
Memory added by this commit:
    user-app: .text +3,676 B  .rodata +284 B  .bss +480 B
    SPDM task: .text +3,648 B  .rodata no change  .bss +448 B
    kernel: no change

    Current totals:
    kernel: .text 151,892 B  .bss 24,536 B
    user-app: .text 58,318 B  .rodata 13,120 B  .bss 61,344 B

SPDM responder validator results:
    DOE: pass: 607, fail: 0
    MCTP: pass: 607, fail: 0

    group 1 (version): pass: 6, fail: 0
    group 2 (capabilities): pass: 58, fail: 0
    group 3 (algorithms): pass: 70, fail: 0
    group 4 (digests): pass: 15, fail: 0
    group 5 (certificate): pass: 113, fail: 0
    group 6 (challenge_auth): pass: 239, fail: 0
    group 7 (measurements): pass: 106, fail: 0

Attestation results:
    EAT: MCTP attestation passed; ocptoken authenticate --show-claims passed
    PCR Quote: MCTP attestation passed; MEASUREMENTS chunking observed via LargeResponse + CHUNK_GET seq 0/1
    PCR Quote measurement signature verification: PASS
