+++ platforms/emulator/runtime/userspace/apps/user/src/spdm/device_measurements/mod.rs
@@ -1,3 +1,5 @@
 // Licensed under the Apache-2.0 license
 
 pub mod ocp_eat;
+#[cfg(feature = "test-mctp-spdm-attestation-pcr-quote")]
+pub mod pcr_quote;
