+++ runtime/userspace/api/caliptra-api-lite/src/device_state.rs
@@ -12,6 +12,7 @@ use crate::ApiAlloc;
 const PCR_VALUE_SIZE: usize = 48;
 const NUM_PCRS: usize = 32;
 const NONCE_SIZE: usize = 32;
+pub const PCR_QUOTE_ECC384_LEN: usize = RSP_SIZE - MBOX_RESP_HEADER_SIZE;
 
 #[repr(C)]
 #[derive(FromBytes, IntoBytes, KnownLayout, Immutable, Unaligned)]
@@ -40,6 +41,45 @@ const RSP_SIZE: usize = MBOX_RESP_HEADER_SIZE
     + 48
     + 48;
 const _: () = assert!(RSP_SIZE == 1848);
+const _: () = assert!(PCR_QUOTE_ECC384_LEN == 1840);
+
+/// Generate a Caliptra ECC PCR quote into `out`, excluding the mailbox response header.
+///
+/// The returned bytes match the non-lite `PcrQuote::pcr_quote(..., with_pqc_sig=false)`
+/// payload shape: PCRs, nonce, reset counters, digest, and ECC signature.
+#[inline(never)]
+pub async fn pcr_quote_ecc384<A: ApiAlloc>(
+    alloc: &A,
+    nonce: Option<&[u8; NONCE_SIZE]>,
+    out: &mut [u8],
+) -> McuResult<usize> {
+    if out.len() < PCR_QUOTE_ECC384_LEN {
+        return Err(INVARIANT);
+    }
+
+    let mut req = alloc.alloc(REQ_SIZE)?;
+    req.fill(0);
+    match nonce {
+        Some(nonce) => req[4..4 + NONCE_SIZE].copy_from_slice(nonce),
+        None => crate::rng_generate(alloc, &mut req[4..4 + NONCE_SIZE]).await?,
+    }
+    let checksum = calc_checksum(CMD_QUOTE_PCRS_ECC384, &req[4..]);
+    req[..4].copy_from_slice(&checksum.to_le_bytes());
+
+    let mut rsp = alloc.alloc(RSP_SIZE)?;
+    let rsp_len = crate::wire::mbox_execute(CMD_QUOTE_PCRS_ECC384, &req, &mut rsp).await?;
+    if rsp_len < RSP_SIZE {
+        return Err(INVARIANT);
+    }
+
+    let nonce_offset = MBOX_RESP_HEADER_SIZE + NUM_PCRS * PCR_VALUE_SIZE;
+    if rsp[nonce_offset..nonce_offset + NONCE_SIZE] != req[4..4 + NONCE_SIZE] {
+        return Err(INVARIANT);
+    }
+
+    out[..PCR_QUOTE_ECC384_LEN].copy_from_slice(&rsp[MBOX_RESP_HEADER_SIZE..RSP_SIZE]);
+    Ok(PCR_QUOTE_ECC384_LEN)
+}
 
 /// Read a single PCR value from the Caliptra QUOTE_PCRS_ECC384 response.
 ///
