+++ runtime/userspace/api/caliptra-api-lite/src/lib.rs
@@ -34,7 +34,7 @@ mod wire;
 
 pub use alloc::ApiAlloc;
 pub use cert::populate_idev_ecc384_cert;
-pub use device_state::get_pcr_value;
+pub use device_state::{get_pcr_value, pcr_quote_ecc384, PCR_QUOTE_ECC384_LEN};
 pub use dpe::{
     dpe_certify_key, dpe_certify_key_pubkey, dpe_get_cert_chain_chunk, dpe_sign_ecc_p384,
     walk_dpe_chain, DpeChainSink, DPE_LABEL_LEN, DPE_MAX_CHUNK_SIZE, DPE_MAX_LEAF_CERT_SIZE,
